🐛(proxy): fix circular references still borking
Sets here weren't working because the objects were consistently re-created because the proxy has to be recursive.
Instead, we temporarily taint the objects as we step through them, removing the taint before they go back out.

This is pretty ugly, but it's the solution I've got right now. Ideally we don't ever do something like this
on the off-chance that somebody has already used the `__tainted__` namespace. Even if we use unique IDs, this is still an issue.